==> 1_Pre-Processing.ipynb <==
import numpy as np
import pandas as pd
pd.set_option('display.max_columns', 30)
pd.set_option('display.max_rows', None)
import matplotlib.pyplot as plt
plt.style.use('ggplot')
import seaborn as sns
import warnings
warnings.filterwarnings('ignore')
import calendar
from tqdm import tqdm

train = pd.read_csv('train.csv')
train.head()

train.shape

test = pd.read_csv('test.csv')
test.head()

test.shape

# Splitting the above data frames into X & y. 
X_train = train.drop('Credit_Score', axis=1)
y_train = train[['Credit_Score']]
y_train['Credit_Score'].replace({'Poor': 0, 'Standard': 1, 'Good': 2}, inplace=True)

X_test = test

X = pd.concat([X_train, X_test], axis=0)
X.reset_index(drop=True, inplace=True)

X.shape

## Pre - Processing

X[X['Customer_ID'] == 'CUS_0xd40']

X[X['Customer_ID'] == 'CUS_0x21b1']

X[X['Customer_ID'] == 'CUS_0xb891']

X.info()

# Checking if each of the 'Customer_ID' are same in number (i.e., 12 as seen from above data frames)
arr = X['Customer_ID'].value_counts().values
all(x == arr[0] for x in arr)

### Confirmed that each 'Customer_ID' has 12 records comprising each Month. We can now make changes and adjustments on a customer level. For 1st step we resolved inconsistencies up until 'Delay_from_due_date' column.

# Creating 'Month' Dictionary to map with Month names with numbers 
month_name = list(calendar.month_name)[1:]
month_num = [i+1 for i, j in enumerate(list(calendar.month_name)[1:])]
month_dict = {}
for i, j in zip(month_name, month_num):
    month_dict[i] = j
#month_dict

X1 = pd.DataFrame(columns=X.columns)

for i in tqdm(X['Customer_ID'].unique()):
    df = X[X['Customer_ID'] == i]

    # Converting 'Month' names to numbers
    df['Month_Num'] = df['Month'].map(month_dict)
    df['Month_Num'] = df['Month_Num'].astype('category')

    # Resolving discrepancies in 'Name' feature
    df['Name'] = df['Name'].mode()[0]

    # 'Age' is tricky. We have resolved this as follows:
    df['Age'] = df['Age'].apply(lambda x: x.replace('_', ''))
    df['Age'] = df['Age'].astype('int64')
    df['Age'] = df['Age'].apply(lambda x: df['Age'].mode()[0] if (x > 94) or (x < 0) else x)

    # 'SSN' and 'Occupation' discrepancies
    df['SSN'] = df['SSN'].mode()[0]
    df['Occupation'] = df['Occupation'].mode()[0]

    # Other Features..
    df['Annual_Income'] = df['Annual_Income'].apply(lambda x: x.replace('_', ''))
    df['Annual_Income'] = df['Annual_Income'].astype('float64')
    df['Annual_Income'] = df['Annual_Income'].mode()[0]

    df['Monthly_Inhand_Salary'].fillna(df['Monthly_Inhand_Salary'].median(), inplace=True)

    df['Num_Bank_Accounts'] = df['Num_Bank_Accounts'].mode()[0]
    df['Num_Credit_Card'] = df['Num_Credit_Card'].mode()[0]
    df['Interest_Rate'] = df['Interest_Rate'].mode()[0]

    df['Num_of_Loan'] = df['Num_of_Loan'].apply(lambda x: x.replace('_', ''))
    df['Num_of_Loan'] = df['Num_of_Loan'].astype('int64')
    df['Num_of_Loan'] = df['Num_of_Loan'].mode()[0]

    df['Type_of_Loan'].fillna('Not Specified', inplace=True)
    df['Delay_from_due_date'] = df['Delay_from_due_date'].apply(lambda x: df['Delay_from_due_date'].mode()[0] if x < 0 else x)

    # Resolving discrepancies in 'Num_of_Delayed_Payment' feature
    df['Num_of_Delayed_Payment'] = df['Num_of_Delayed_Payment'].astype(str)
    df['Num_of_Delayed_Payment'] = df['Num_of_Delayed_Payment'].apply(lambda x: x.replace('_', ''))
    df['Num_of_Delayed_Payment'].replace('nan', np.NaN, inplace=True)
    df['Num_of_Delayed_Payment'].fillna(df['Num_of_Delayed_Payment'].mode()[0], inplace=True)
    df['Num_of_Delayed_Payment'] = df['Num_of_Delayed_Payment'].astype('int64')
    df['Num_of_Delayed_Payment'] = df['Num_of_Delayed_Payment'].apply(lambda x: df['Num_of_Delayed_Payment'].mode()[0] if (x > 25) else x)
    df['Num_of_Delayed_Payment'] = df['Num_of_Delayed_Payment'].apply(lambda x: np.abs(int(x)) if (x < 0) else x)
    
    X1 = pd.concat([X1, df], axis=0)

X = X1.sort_index()
X['Month'] = X['Month_Num']
X.drop('Month_Num', axis=1, inplace=True)

X.shape

X.info()

X1 = pd.DataFrame(columns=X.columns)

for i in tqdm(X['Customer_ID'].unique()):
    df = X[X['Customer_ID'] == i]

    # 'Changed_Credit_Limit' feature
    df['Changed_Credit_Limit'].replace('_', np.NaN, inplace=True)
    df['Changed_Credit_Limit'] = df['Changed_Credit_Limit'].astype('float64')
    df['Changed_Credit_Limit'] = df['Changed_Credit_Limit'].apply(lambda x: df['Changed_Credit_Limit'].mode()[0] if (x < 0) else x)
    df['Changed_Credit_Limit'].fillna(df['Changed_Credit_Limit'].mode()[0], inplace=True)
    df['Changed_Credit_Limit'] = df['Changed_Credit_Limit'].apply(lambda x: round(x, 2))

    # 'Num_Credit_Inquiries' feature
    df['Num_Credit_Inquiries'] = df['Num_Credit_Inquiries'].apply(lambda x: df['Num_Credit_Inquiries'].mode()[0] if (x > 20) else x)
    df['Num_Credit_Inquiries'].fillna(df['Num_Credit_Inquiries'].mode()[0], inplace=True)
    df['Num_Credit_Inquiries'] = df['Num_Credit_Inquiries'].astype('int64')

    # Addressing discrepancies in 'Credit Mix' feature
    df['Credit_Mix'].replace('_', df['Credit_Mix'].mode()[0], inplace=True)

    # 'Outstanding_Debt' feature
    df['Outstanding_Debt'] = df['Outstanding_Debt'].apply(lambda x: x.replace('_', ''))
    df['Outstanding_Debt'] = df['Outstanding_Debt'].astype('float64')

    # Dropping 'Credit_History_Age' feature entirely
    df.drop('Credit_History_Age', axis=1, inplace=True)

    # 'Amount_invested_monthly' feature
    df['Amount_invested_monthly'].replace('__10000__', np.NaN, inplace=True)
    df['Amount_invested_monthly'] = df['Amount_invested_monthly'].astype('float64')
    df['Amount_invested_monthly'].fillna(df['Amount_invested_monthly'].median(), inplace=True)

    # 'Payment_Behaviour' feature
    df['Payment_Behaviour'].replace('!@9#%8', df['Payment_Behaviour'].mode()[0], inplace=True)

    # 'Monthly_Balance' feature
    df['Monthly_Balance'].replace('__-333333333333333333333333333__', np.NaN, inplace=True)
    df['Monthly_Balance'] = df['Monthly_Balance'].astype('float64')
    df['Monthly_Balance'].fillna(df['Monthly_Balance'].median(), inplace=True)

    X1 = pd.concat([X1, df], axis=0)

X = X1.sort_index()

# Again assigning data types correctly for final time
X['Age'] = X['Age'].astype('int64')
X['Annual_Income'] = X['Annual_Income'].astype('float64')
X['Num_Bank_Accounts'] = X['Num_Bank_Accounts'].astype('int64')
X['Num_Credit_Card'] = X['Num_Credit_Card'].astype('int64')
X['Interest_Rate'] = X['Interest_Rate'].astype('int64')
X['Num_of_Loan'] = X['Num_of_Loan'].astype('int64')
X['Delay_from_due_date'] = X['Delay_from_due_date'].astype('int64')
X['Num_of_Delayed_Payment'] = X['Num_of_Delayed_Payment'].astype('int64')
X['Num_Credit_Inquiries'] = X['Num_Credit_Inquiries'].astype('int64')

X.shape

X.head()

X.info()

X.drop(columns=['Customer_ID' ,'Name', 'SSN', 'Credit_History_Age'], inplace=True)

### Looking into Categorical columns

for i in X.select_dtypes('object').columns:
    print(f"No. of Unique values in {i} is {X[i].unique().shape[0]}")

### Dropping 'Type_of_Loan' feature as it has too many different values which may lead to overfitting and complexity of model.

X.drop('Type_of_Loan', axis=1, inplace=True)

### Performing Label Encoding on remaining category columns (Except for 'ID'). 

for i in X.select_dtypes('object').columns[1:]:
    X[i] = pd.Categorical(X[i]).codes

for i in X.columns:
    if X[i].dtype == 'int8':
        X[i] = X[i].astype('category')

X.info()

category_col = list(X.select_dtypes('category').columns)
category_col

### Dividing Train and Test data

cleaned_train = pd.concat([X[X.index < 100000], y_train], axis=1)

cleaned_train.head()

cleaned_test = X[X.index >= 100000].reset_index(drop=True)

cleaned_test.head()

### Converting them to CSV Files

cleaned_train.to_csv('Cleaned Train.csv', index=False)

cleaned_test.to_csv('Cleaned Test.csv', index=False)

## END OF PRE-PROCESSING.
