==> 3_Model_Training_1.ipynb <==
import numpy as np
import pandas as pd
pd.set_option('display.max_columns', 30)
pd.set_option('display.max_rows', None)
import matplotlib.pyplot as plt
plt.style.use('ggplot')
import seaborn as sns
import warnings
warnings.filterwarnings('ignore')

from scipy.stats import zscore
from sklearn.model_selection import train_test_split, RandomizedSearchCV, cross_val_score, ShuffleSplit
from sklearn.ensemble import GradientBoostingClassifier, RandomForestClassifier
from xgboost import XGBClassifier

# Alter the file path as you see fit
df = pd.read_csv('Cleaned Train.csv')
df.head()

df.set_index('ID', drop=True, inplace=True)

X = df.drop('Credit_Score', axis=1)
y = df[['Credit_Score']]
y1 = y.copy()
y['Credit_Score'] = y['Credit_Score'].astype('category')

y.info()

y.head()

y['Credit_Score'].value_counts(normalize=True)*100

# Converting Cateogry columns to 'category' data type
cat_col = ['Month', 'Occupation', 'Credit_Mix', 'Payment_of_Min_Amount', 'Payment_Behaviour']
for i in cat_col:
    X[i] = X[i].astype('category')

X.info()

X.describe().T

### Let's look at Cross Validation accuracy scores for various Classifiers without Hypertuning any parameters

print('Random Forest =',np.mean(cross_val_score(RandomForestClassifier(max_depth=7, min_samples_leaf=3), X, y, scoring='accuracy',)))
print('Gradient Boosting =',np.mean(cross_val_score(GradientBoostingClassifier(max_depth=7, min_samples_leaf=3), X, y, cv=3, scoring='accuracy')))
print('XG Boosting =',np.mean(cross_val_score(XGBClassifier(enable_categorical=True), X, y1, scoring='accuracy')))

### As we can see, simple classifiers aren't giving good accuracies on data. Let's proceed to hypertune the parameters.

### Training the models below with all the 100,000 records was computationally very intensive so we have splitted this Train data again to train-test datasets. We will be hypertuning the models with this sub-train set.

# 85:15 split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.15, random_state=42)

## 1. Random Forest

RandomForestClassifier().get_params()

shuffle_split = ShuffleSplit(n_splits=1, test_size=0.2, random_state=42)

param_grid = {'n_estimators': [100, 150, 200],
              'criterion': ['gini', 'entropy'],
              'max_depth': [3, 5, 7],
              'max_features': ['sqrt', len(X.columns)],
              'min_samples_leaf': [6, 7, 8, 9, 10],
              'oob_score': [True, False]}

rand_search2 = RandomizedSearchCV(RandomForestClassifier(), param_grid, scoring='accuracy', n_iter=100, verbose=2, cv=shuffle_split, random_state=42)
rand_search2.fit(X_train, y_train)

rand_search2.best_params_

# Defining a dataframe to add all accuracies of different models
a = pd.DataFrame(columns=['Model', 'Performance on Sub-Test set', 'Accuracy of CV on Entire Train set'])

# Let's look at performance on test set
model_rf = rand_search2.best_estimator_
model_rf.fit(X_train, y_train)
a2 = model_rf.score(X_test, y_test)

# Let's look at Cross Validation score
model_rf = rand_search2.best_estimator_
b2 = cross_val_score(model_rf, X, y, cv=5, scoring='accuracy')
b2

print('Mean Accuracy from Cross Validation =',np.mean(b2))
a.loc[len(a.index)] = ['Random Forest', round(a2, 2), round(np.mean(b2), 2)]

## 2. Gradient Boosting

GradientBoostingClassifier().get_params()

param_grid = {'n_estimators': [100, 150, 200],
              'criterion': ['log_loss', 'exponential'],
              'max_depth': [5, 7, 9],
              'max_features': ['sqrt', len(X.columns)],
              'min_samples_leaf': [7, 8, 9, 10],
              'criterion': ['friedman_mse', 'squared_error'],
              'validation_fraction': [0.1, 0.2, 0.3]}

rand_search4 = RandomizedSearchCV(GradientBoostingClassifier(), param_grid, scoring='accuracy', n_iter=100, cv=shuffle_split, verbose=2, random_state=42)
rand_search4.fit(X_train, y_train)

rand_search4.best_params_

# Let's look at performance on test set
model_gb = rand_search4.best_estimator_
model_gb.fit(X_train, y_train)
a4 = model_gb.score(X_test, y_test)

# Let's look at Cross Validation score
model_gb = rand_search4.best_estimator_
b4 = cross_val_score(model_gb, X, y, cv=5, scoring='accuracy')
b4

print('Mean Accuracy from Cross Validation =',np.mean(b4))
a.loc[len(a.index)] = ['Gradient Boosting', round(a4, 2), round(np.mean(b4), 2)]

a

# THE END
