==> 4_Model_Training_2.ipynb <==
import numpy as np
import pandas as pd
pd.set_option('display.max_columns', 30)
pd.set_option('display.max_rows', None)
import matplotlib.pyplot as plt
plt.style.use('ggplot')
import seaborn as sns
import warnings
warnings.filterwarnings('ignore')

from scipy.stats import zscore
from sklearn.model_selection import train_test_split, GridSearchCV, RandomizedSearchCV, cross_val_score, ShuffleSplit
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import AdaBoostClassifier

# Alter the file path as you see fit
df = pd.read_csv('Cleaned Train.csv')
df.head()

df.set_index('ID', drop=True, inplace=True)

X = df.drop('Credit_Score', axis=1)
y = df[['Credit_Score']]
y1 = y.copy()
y['Credit_Score'] = y['Credit_Score'].astype('category')

y.info()

y.head()

y['Credit_Score'].value_counts(normalize=True)*100

# Converting Cateogry columns to 'category' data type
cat_col = ['Month', 'Occupation', 'Credit_Mix', 'Payment_of_Min_Amount', 'Payment_Behaviour']
for i in cat_col:
    X[i] = X[i].astype('category')

X.info()

X.describe().T

### Let's look at Cross Validation accuracy scores for various Classifiers without Hypertuning any parameters

print('Simple Decision Tree =',np.mean(cross_val_score(DecisionTreeClassifier(max_depth=7, min_samples_leaf=3), X, y, scoring='accuracy')))
print('Ada Boosting =',np.mean(cross_val_score(AdaBoostClassifier(n_estimators=100), X, y, scoring='accuracy')))

### As we can see, simple classifiers aren't giving good accuracies on data. Let's proceed to hypertune the parameters.

### Training the models below with all the 100,000 records was computationally very intensive so we have splitted this Train data again to train-test datasets. We will be hypertuning the models with this sub-train set.

# 85:15 split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.15, random_state=42)

## 1. Decision Tree

DecisionTreeClassifier().get_params()

param_grid = {'criterion': ['gini', 'log_loss', 'entropy'],
              'max_depth': [3, 5, 7],
              'max_features': ['sqrt', len(X.columns)],
              'min_samples_leaf': [6, 7, 8, 9, 10]}

grid_search1 = GridSearchCV(DecisionTreeClassifier(), param_grid, scoring='accuracy')
grid_search1.fit(X_train, y_train)

grid_search1.best_params_

# Defining a dataframe to add all accuracies of different models
a = pd.DataFrame(columns=['Model', 'Performance on Sub-Test set', 'Accuracy of CV on Entire Train set'])

# Let's look at performance on test set
model_dt = grid_search1.best_estimator_
model_dt.fit(X_train, y_train)
a1 = model_dt.score(X_test, y_test)

# Let's look at Cross Validation score for the entire Train set i.e., not only the sub-test but entire dataset
model_dt = grid_search1.best_estimator_
b1 = cross_val_score(model_dt, X, y, cv=5, scoring='accuracy')
b1

print('Mean Accuracy from Cross Validation =',np.mean(b1))
a.loc[len(a.index)] = ['Decision Tree', round(a1, 2), round(np.mean(b1), 2)]

## 2. Ada Boosting

AdaBoostClassifier().get_params()

shuffle_split = ShuffleSplit(n_splits=1, test_size=0.2, random_state=42)

param_grid = {'estimator': [DecisionTreeClassifier(max_depth=5, min_samples_leaf=3), DecisionTreeClassifier(max_depth=7, min_samples_leaf=5), model_dt],
              'n_estimators': [100, 150, 200, 250],
              'learning_rate': [1, 2]}

rand_search3 = RandomizedSearchCV(AdaBoostClassifier(), param_grid, scoring='accuracy', n_iter=30, cv=shuffle_split, random_state=42)
rand_search3.fit(X_train, y_train)

rand_search3.best_params_

# Let's look at performance on test set
model_ada = rand_search3.best_estimator_
model_ada.fit(X_train, y_train)
a3 = model_ada.score(X_test, y_test)

# Let's look at Cross Validation score
model_ada = rand_search3.best_estimator_
b3 = cross_val_score(model_ada, X, y, cv=5, scoring='accuracy')
b3

print('Mean Accuracy from Cross Validation =',np.mean(b3))
a.loc[len(a.index)] = ['Ada Boosting', round(a3, 2), round(np.mean(b3), 2)]

a

# THE END
