==> 5_Model_Training_3.ipynb <==
import numpy as np
import pandas as pd
pd.set_option('display.max_columns', 30)
pd.set_option('display.max_rows', None)
import matplotlib.pyplot as plt
plt.style.use('ggplot')
import seaborn as sns
import warnings
warnings.filterwarnings('ignore')

from scipy.stats import zscore
from sklearn.model_selection import GridSearchCV, RandomizedSearchCV, cross_val_score
from sklearn.neighbors import KNeighborsClassifier
from xgboost import XGBClassifier

# Alter the file path as you see fit
df = pd.read_csv('Cleaned Train.csv')
df.head()

df.set_index('ID', drop=True, inplace=True)

X = df.drop('Credit_Score', axis=1)
y = df[['Credit_Score']]
y1 = y.copy()
y['Credit_Score'] = y['Credit_Score'].astype('category')

X.info()

y.info()

y['Credit_Score'].value_counts(normalize=True)*100

# Scaling X
X_scaled = X.apply(zscore)

### Let's look at Cross Validation accuracy scores for various Classifiers without Hypertuning any parameters

print('KNN =',np.mean(cross_val_score(KNeighborsClassifier(), X_scaled, y, scoring='accuracy')))
print('XG Boosting =',np.mean(cross_val_score(XGBClassifier(enable_categorical=True), X, y1, scoring='accuracy')))

### As we can see, simple classifiers aren't giving good accuracies on data. Let's proceed to hypertune the parameters.

## 1. KNN Classification

KNeighborsClassifier().get_params()

param_grid = {'n_neighbors': [3, 5, 7],
              'weights': ['uniform', 'distance'],
              'algorithm': ['auto', 'brute']}

grid_search1 = GridSearchCV(KNeighborsClassifier(), param_grid, scoring='accuracy')
grid_search1.fit(X_scaled, y)

grid_search1.best_params_

# Let's look at Cross Validation score now
model_dt = grid_search1.best_estimator_
a1 = cross_val_score(model_dt, X_scaled, y, cv=5, scoring='accuracy')
a1

print('Mean Accuracy from Cross Validation =',np.mean(a1))

## 2. XG Boosting

#XGBClassifier().get_params()

param_grid = {
    'n_estimators': [100, 150, 200, 250, 300],  # Number of boosting rounds
    'learning_rate': [0.01, 0.1, 0.2],  # Rate at which the model's weights are updated
    'max_depth': [3, 5, 7],  # Maximum depth of each tree
    'min_child_weight': [1, 3, 5],  # Minimum sum of instance weight (hessian) needed in a child
    'subsample': [0.6, 0.8, 1.0],  # Fraction of samples used for fitting each tree
    'colsample_bytree': [0.6, 0.8, 1.0],  # Fraction of features used for fitting each tree
    'gamma': [0, 0.1, 0.2],  # Minimum loss reduction required to make a further partition on a leaf node
    'reg_alpha': [0, 0.1, 0.5],  # L1 regularization term on weights
    'reg_lambda': [0, 0.1, 0.5],  # L2 regularization term on weights
    'scale_pos_weight': [1, 2, 3]  # Control the balance of positive and negative weights
}

rand_search2 = RandomizedSearchCV(XGBClassifier(), param_grid, scoring='accuracy', n_iter=100, random_state=42)
rand_search2.fit(X, y1)

rand_search2.best_params_

# Let's look at Cross Validation score now
model_xgb = rand_search2.best_estimator_
a2 = cross_val_score(model_xgb, X, y1, cv=5, scoring='accuracy')
a2

print('Mean Accuracy from Cross Validation =',np.mean(a2))
