==> 6_Final_Prediction.ipynb <==
import numpy as np
import pandas as pd
pd.set_option('display.max_columns', 30)
pd.set_option('display.max_rows', None)
import warnings
warnings.filterwarnings('ignore')

from sklearn.ensemble import GradientBoostingClassifier

df_train = pd.read_csv('Cleaned Train.csv')
df_train.head()

df_train.set_index('ID', drop=True, inplace=True)

X_train = df_train.drop('Credit_Score', axis=1)
y_train = df_train[['Credit_Score']]
y_train['Credit_Score'] = y_train['Credit_Score'].astype('category')

X_test = pd.read_csv('Cleaned Test.csv', index_col='ID')
X_test.head()

# Training Gradient Boost Model with best parameters deduced from hypertuning
model = GradientBoostingClassifier(validation_fraction=0.1, n_estimators=200, min_samples_leaf=8, 
                                   max_features=21, max_depth=9, criterion='friedman_mse')

model.fit(X_train, y_train)

pred = model.predict(X_test)
pred

output = pd.DataFrame({'ID': X_test.index, 'Predicted_values': pred})
output.head(10)

output.to_csv('Output.csv', index=True)
