==> 7_Additional_Stacking_Method.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from tqdm import tqdm

from scipy.stats import zscore
from sklearn.model_selection import train_test_split, KFold
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier, GradientBoostingClassifier
from xgboost import XGBClassifier
from sklearn.metrics import confusion_matrix
import warnings
warnings.filterwarnings('ignore')

df = pd.read_csv('/content/drive/MyDrive/Colab Notebooks/Supervised Learning Mini Project/Cleaned Train.csv')
df.head()

df.set_index('ID', drop=True, inplace=True)

X = df.drop('Credit_Score', axis=1)
y = df[['Credit_Score']]
y['Credit_Score'] = y['Credit_Score'].astype('category')

# Converting Cateogry columns to 'category' data type
cat_col = ['Month', 'Occupation', 'Credit_Mix', 'Payment_of_Min_Amount', 'Payment_Behaviour']
for i in cat_col:
    X[i] = X[i].astype('category')

X.info()

X_train_final, X_test_final, y_train_final, y_test_final = train_test_split(X, y, test_size=0.15, random_state=42)
X_train_final.reset_index(drop=True, inplace=True)
X_test_final.reset_index(drop=True, inplace=True)
y_train_final.reset_index(drop=True, inplace=True)
y_test_final.reset_index(drop=True, inplace=True)

estimators = [DecisionTreeClassifier(criterion='gini', max_depth=7, max_features=21, min_samples_leaf=8),
              RandomForestClassifier(oob_score=True, n_estimators=150, min_samples_leaf=8, max_features=21, max_depth=7, criterion='gini'),
              GradientBoostingClassifier(validation_fraction=0.2, n_estimators=200, min_samples_leaf=7, max_features=21, max_depth=9,
                                         criterion='squared_error')]

all_predictions = pd.DataFrame()
trained_models = []
kf = KFold(10, shuffle=True)
checker = 0
for i in tqdm(range(len(estimators))):
    for train_index, test_index in kf.split(X_train_final):
        X_train, X_test = X_train_final.loc[train_index], X_train_final.loc[test_index]
        y_train, y_test = y_train_final.loc[train_index], y_train_final.loc[test_index]

        model = estimators[i]
        m = model.fit(X_train, y_train)
        trained_models.append(m)
        checker += 1
        print(checker, end='')

        model_pred_train = pd.DataFrame(model.predict(X_train), index=train_index)
        model_pred_test = pd.DataFrame(model.predict(X_test), index=test_index)

        a = pd.concat([model_pred_train, model_pred_test], axis=0)
        all_predictions = pd.concat([all_predictions, a.sort_index()], axis=1)

all_predictions.columns = list(range(30))

all_predictions

trained_models[0]

y_train_final['Credit_Score'] = y_train_final['Credit_Score'].astype('int64')
y_test_final['Credit_Score'] = y_test_final['Credit_Score'].astype('int64')

# Main Model
model_main = XGBClassifier(subsample=0.6, scale_pos_weight=1, reg_lambda=0, reg_alpha=0.1,
                           n_estimators=150, min_child_weight=1, max_depth=7, learning_rate=0.01, gamma=0, colsample_bytree=1.0)
model_main.fit(all_predictions, y_train_final)

X_test_final.head()

test_predictions = pd.DataFrame(columns=list(range(30)))

dummy = []
for i in range(len(X_test_final)):
    for j in range(len(trained_models)):
        dummy.append(trained_models[j].predict(X_test_final[X_test_final.index == i])[0])
    test_predictions.loc[len(test_predictions.index)] = dummy
    dummy = []

test_predictions.head(20)

y_pred = pd.DataFrame(model_main.predict(test_predictions), columns=['Credit_Score'], index=y_test_final.index)

y_test_final

print(y_pred.shape)
print(y_test_final.shape)

model_main.score(test_predictions, y_test_final)

confusion_matrix(y_test_final, y_pred)
